+++ obp-api/src/main/scala/code/api/ResourceDocs1_4_0/SwaggerJSONFactory.scala
@@ -19,7 +19,7 @@ import com.openbankproject.commons.model.JsonFieldReName
 import net.liftweb.util.StringHelpers
 
 import scala.collection.mutable.ListBuffer
-import code.api.v3_1_0.ListResult
+import com.openbankproject.commons.model.ListResult
 import code.util.Helper.MdcLoggable
 import com.openbankproject.commons.util.JsonUtils
 import net.liftweb.common.{EmptyBox, Full}
