+++ obp-api/src/test/scala/code/api/v4_0_0/CustomerAttributesTest.scala
@@ -6,7 +6,8 @@ import code.api.util.ApiRole._
 import code.api.util.ErrorMessages.{CustomerAttributeNotFound, UserHasMissingRoles, UserNotLoggedIn}
 import code.api.util.ExampleValue.{customerAttributeValueExample,customerAttributeNameExample}
 import code.api.v3_0_0.CustomerAttributeResponseJsonV300
-import code.api.v3_1_0.{CustomerWithAttributesJsonV310, ListResult}
+import code.api.v3_1_0.CustomerWithAttributesJsonV310
+import com.openbankproject.commons.model.ListResult
 import code.api.v4_0_0.OBPAPI4_0_0.Implementations4_0_0
 import code.entitlement.Entitlement
 import com.github.dwickern.macros.NameOf.nameOf
