+++ obp-commons/pom.xml
@@ -31,6 +31,10 @@
             <artifactId>lift-util_${scala.version}</artifactId>
         </dependency>
         <dependency>
+            <groupId>net.liftweb</groupId>
+            <artifactId>lift-mapper_${scala.version}</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.scala-lang</groupId>
             <artifactId>scala-reflect</artifactId>
             <scope>compile</scope>
